Tęsiamas programos portinimas į OOP

+++ Studentas.h
@@ -6,15 +6,15 @@ class Studentas {
 private:
     std::string name_;
     std::string surname_;
-    double examGrade_;
-    std::vector<double> homeworkGrades_;
+    int examGrade_;
+    std::vector<int> homeworkGrades_;
 
 public:
     Studentas() : examGrade_(0.0) {}
     Studentas(const std::string& name,
               const std::string& surname,
-              double examGrade,
-              const std::vector<double>& homeworkGrades)
+              int examGrade,
+              const std::vector<int>& homeworkGrades)
         : name_(name),
           surname_(surname),
           examGrade_(examGrade),
@@ -28,13 +28,13 @@ public:
     }
     inline std::string getName() const { return name_; }
     inline std::string getSurname() const { return surname_; }
-    inline double getExamGrade() const { return examGrade_; }
-    inline std::vector<double> getHomeworkGrades() const { return homeworkGrades_; }
-    double calculateFinalGrade(double (*calcFunc)(const std::vector<double>&)) const;
+    inline int getExamGrade() const { return examGrade_; }
+    inline std::vector<int> getHomeworkGrades() const { return homeworkGrades_; }
+    double calculateFinalGrade(double (*calcFunc)(const std::vector<int>&)) const;
     std::istream& readStudent(std::istream&);
     void setName(const std::string& name) { name_ = name; }
     void setSurname(const std::string& surname) { surname_ = surname; }
-    void setExamGrade(double examGrade) { examGrade_ = examGrade; }
-    void setHomeworkGrades(const std::vector<double>& homeworkGrades) {homeworkGrades_ = homeworkGrades;}
-    void addHomeworkGrade(double grade) {homeworkGrades_.push_back(grade);}
+    void setExamGrade(int examGrade) { examGrade_ = examGrade; }
+    void setHomeworkGrades(const std::vector<int>& homeworkGrades) {homeworkGrades_ = homeworkGrades;}
+    void addHomeworkGrade(int grade) {homeworkGrades_.push_back(grade);}
 };
